+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/points.ts
@@ -8,11 +8,11 @@
 
 import { RgbObject } from '../../../../common/color_library_wrappers';
 import { SeriesKey } from '../../../../common/series_id';
-import { Circle, Stroke, Fill, Rect } from '../../../../geoms/types';
+import { Circle, Fill, Rect, Stroke } from '../../../../geoms/types';
 import { Rotation } from '../../../../utils/common';
 import { Dimensions } from '../../../../utils/dimensions';
 import { PointGeometry } from '../../../../utils/geometry';
-import { PointStyle, GeometryStateStyle, PointShape } from '../../../../utils/themes/theme';
+import { GeometryStateStyle } from '../../../../utils/themes/theme';
 import { renderShape } from './primitives/shapes';
 import { withPanelTransform } from './utils/panel_transform';
 
@@ -23,26 +23,14 @@ import { withPanelTransform } from './utils/panel_transform';
  */
 export function renderPoints(ctx: CanvasRenderingContext2D, points: PointGeometry[], { opacity }: GeometryStateStyle) {
   points
-    .map<[Circle, Fill, Stroke, PointShape]>(({ x, y, radius, transform, style }) => {
-      const fill: Fill = {
-        color: applyOpacity(style.fill.color, opacity),
-      };
-
-      const stroke: Stroke = {
-        ...style.stroke,
-        color: applyOpacity(style.stroke.color, opacity),
-      };
-
-      const coordinates: Circle = {
-        x: x + transform.x,
-        y: y + transform.y,
-        radius,
-      };
-
-      return [coordinates, fill, stroke, style.shape];
-    })
-    .sort(([{ radius: a }], [{ radius: b }]) => b - a)
-    .forEach(([coordinates, fill, stroke, shape]) => renderShape(ctx, shape, coordinates, fill, stroke));
+    .slice()
+    .sort(({ radius: a }, { radius: b }) => b - a)
+    .forEach(({ x, y, radius, transform, style }) => {
+      const coordinates = { x: x + transform.x, y: y + transform.y, radius };
+      const fill = { color: applyOpacity(style.fill.color, opacity) };
+      const stroke = { ...style.stroke, color: applyOpacity(style.stroke.color, opacity) };
+      renderShape(ctx, style.shape, coordinates, fill, stroke);
+    });
 }
 
 /**
@@ -53,7 +41,6 @@ export function renderPoints(ctx: CanvasRenderingContext2D, points: PointGeometr
 export function renderPointGroup(
   ctx: CanvasRenderingContext2D,
   points: PointGeometry[],
-  themeStyles: Record<SeriesKey, PointStyle>,
   geometryStateStyles: Record<SeriesKey, GeometryStateStyle>,
   rotation: Rotation,
   renderingArea: Dimensions,
@@ -61,45 +48,18 @@ export function renderPointGroup(
   shouldClip: boolean,
 ) {
   points
-    .map<[Circle, Fill, Stroke, Dimensions, PointShape]>(
-      ({ x, y, radius, transform, style, seriesIdentifier: { key }, panel }) => {
-        const { opacity } = geometryStateStyles[key];
-        const fill: Fill = {
-          color: applyOpacity(style.fill.color, opacity),
-        };
-
-        const stroke: Stroke = {
-          ...style.stroke,
-          color: applyOpacity(style.stroke.color, opacity),
-        };
-
-        const coordinates: Circle = {
-          x: x + transform.x,
-          y,
-          radius,
-        };
-
-        return [coordinates, fill, stroke, panel, style.shape];
-      },
-    )
-    .sort(([{ radius: a }], [{ radius: b }]) => b - a)
-    .forEach(([coordinates, fill, stroke, panel, shape]) => {
-      withPanelTransform(
-        ctx,
-        panel,
-        rotation,
-        renderingArea,
-        (ctx) => {
-          renderShape(ctx, shape, coordinates, fill, stroke);
-        },
-        { area: clippings, shouldClip },
-      );
+    .slice()
+    .sort(({ radius: a }, { radius: b }) => b - a)
+    .forEach(({ x, y, radius, transform, style, seriesIdentifier: { key }, panel }) => {
+      const { opacity } = geometryStateStyles[key];
+      const fill: Fill = { color: applyOpacity(style.fill.color, opacity) };
+      const stroke: Stroke = { ...style.stroke, color: applyOpacity(style.stroke.color, opacity) };
+      const coordinates: Circle = { x: x + transform.x, y, radius };
+      const renderer = () => renderShape(ctx, style.shape, coordinates, fill, stroke);
+      withPanelTransform(ctx, panel, rotation, renderingArea, renderer, { area: clippings, shouldClip });
     });
 }
 
 function applyOpacity(color: RgbObject, opacity: number): RgbObject {
-  return {
-    ...color,
-    opacity: color.opacity * opacity,
-  };
+  return { ...color, opacity: color.opacity * opacity };
 }